requirements.txt
# Django
Django==5.0.3

# Django REST Framework
djangorestframework==3.15.0

# Channels
channels[daphne]==4.0.0
channels-redis==4.2.0

# Connector to Redis Server
redis==5.0.3

# Celery
celery==5.3.6
gevent==24.2.1

# Utils
everett==3.3.0
